pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.2",
    "aiosqlite>=0.21.0",
    "anthropic>=0.75.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.122.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "markdown>=3.10",
    "matplotlib>=3.10.7",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "psutil>=7.1.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.5",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "weasyprint>=66.0",
    "websockets>=15.0.1",
]
